refactor(scanner): Move `mergeScanResults()` to the `model` module
Enable re-use in an upcoming change.
